requirements.txt
requests==2.26.0
srsly==2.4.2
py-processors== 3.2.2
networkx==2.6.3